refactor: moved sender_for_bundle interaction to own function.
